# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'BVMS Admin'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'https://bvms.online'
paths:
...
  /api/forgot-password:
    post:
      summary: 'Forgot Password'
      operationId: forgotPassword
      description: "Send a reset password link to the user's email."
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Reset password link sent to your email'
                properties:
                  message:
                    type: string
                    example: 'Reset password link sent to your email'
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'User not found'
                properties:
                  error:
                    type: string
                    example: 'User not found'
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: "The user's email address."
                  example: user@example.com
                  nullable: false
              required:
                - email
      security: []
...